Rename alt-file-parameter → file-parameters
