+++ .github/CODEOWNERS
@@ -1,1 +1,1 @@
-* @jenkinsci/alt-file-parameter-plugin-plugin-developers
+* @jenkinsci/file-parameters-plugin-plugin-developers
