+++ src/test/java/io/jenkins/plugins/file_parameters/AbstractFileParameterDefinitionTest.java
@@ -22,7 +22,7 @@
  * THE SOFTWARE.
  */
 
-package io.jenkins.plugins.alt_file_parameter;
+package io.jenkins.plugins.file_parameters;
 
 import hudson.cli.CLICommandInvoker;
 import hudson.model.ParametersDefinitionProperty;
